refactor: [#31] :building_construction: Update architecture
