+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.color.ColorsActivity" />
-        <activity android:name=".ResultActivity" />
-        <activity android:name=".QuestionActivity" />
+        <activity android:name=".color.ColorsActivity" />
+        <activity android:name=".result.ResultActivity" />
+        <activity android:name=".question.QuestionActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
