+++ app/src/main/java/fr/esgi/rpa/cgg/MainActivity.kt
@@ -3,7 +3,8 @@ package fr.esgi.rpa.cgg
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import fr.esgi.rpa.cgg.ui.color.ColorsActivity
+import fr.esgi.rpa.cgg.color.ColorsActivity
+import fr.esgi.rpa.cgg.question.QuestionActivity
 import kotlinx.android.synthetic.main.activity_main.*
 
 // TODO: refactor navigation
