+++ app/src/main/java/fr/esgi/rpa/cgg/color/Color.kt
@@ -1,3 +1,3 @@
-package fr.esgi.rpa.cgg.data
+package fr.esgi.rpa.cgg.color
 
 data class Color(val name: String, val lightValue: String, val darkValue: String)
